# api-docs.yaml
openapi: "3.0.3"
info:
  title: "CRM System API"
  version: "1.0.0"
  description: "Sistema de gerenciamento de relacionamento com clientes - API completa para operações CRUD e relatórios de dashboard"
  contact:
    name: "CRM System Support"
    email: "support@crm-system.com"
  license:
    name: "ISC"
    url: "https://opensource.org/licenses/ISC"
servers:
  - url: "http://localhost:3000/api"
    description: "Development server"
  - url: "https://api.crm-system.com/api"
    description: "Production server"
components:
  schemas:
...
    CreateBusinessProposalRequest:
      type: "object"
      properties:
        business:
          type: "object"
          properties:
            id:
              type: "string"
              format: "uuid"
          required:
            - id
        responsible:
          type: "object"
          properties:
            id:
              type: "string"
              format: "uuid"
          required:
            - id
        title:
          type: "string"
          minLength: 1
        status:
          type: "string"
        date:
          type: "string"
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        value:
          type: "number"
          minimum: 0
          exclusiveMinimum: true
        content:
          type: "string"
        themeColor:
          type: "string"
          pattern: "^#[0-9A-Fa-f]{6}$"
        termsAndConditions:
          type: "string"
        showUnitPrices:
          type: "boolean"
      required:
        - business
        - responsible
        - title
        - date
        - value
      description: "Schema for creating a new business proposal"
      example:
        business:
          id: "123e4567-e89b-12d3-a456-426614174000"
        responsible:
          id: "456e7890-e89b-12d3-a456-426614174000"
        title: "Software Implementation Proposal"
        status: "Rascunho"
        date: "2024-01-15"
        value: 50000
        content: "Detailed proposal content..."
        themeColor: "#007bff"
        termsAndConditions: "Standard terms and conditions..."
        showUnitPrices: true
    UpdateBusinessProposalRequest:
      type: "object"
      properties:
        business:
          type: "object"
          properties:
            id:
              type: "string"
              format: "uuid"
          required:
            - id
        responsible:
          type: "object"
          properties:
            id:
              type: "string"
              format: "uuid"
          required:
            - id
        title:
          type: "string"
          minLength: 1
        status:
          type: "string"
        date:
          type: "string"
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        value:
          type: "number"
          minimum: 0
          exclusiveMinimum: true
        content:
          anyOf:
            - type: "string"
            - nullable: true
            - nullable: true
        themeColor:
          anyOf:
            - type: "string"
              pattern: "^#[0-9A-Fa-f]{6}$"
            - nullable: true
            - nullable: true
        termsAndConditions:
          anyOf:
            - type: "string"
            - nullable: true
            - nullable: true
        showUnitPrices:
          anyOf:
            - type: "boolean"
            - nullable: true
            - nullable: true
      description: "Schema for updating an existing business proposal"
      example:
        status: "Em Revisão"
        value: 55000
        content: "Updated proposal content..."
...
    BusinessProposal:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        business:
          type: "object"
          properties:
            id:
              type: "string"
              format: "uuid"
          required:
            - id
        responsible:
          type: "object"
          properties:
            id:
              type: "string"
              format: "uuid"
          required:
            - id
        title:
          type: "string"
        status:
          type: "string"
        date:
          type: "string"
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        value:
          type: "number"
          minimum: 0
          exclusiveMinimum: true
        content:
          type: "string"
        themeColor:
          type: "string"
          pattern: "^#[0-9A-Fa-f]{6}$"
        termsAndConditions:
          type: "string"
        showUnitPrices:
          type: "boolean"
        sendMessage:
          type: "string"
          description: "Optional message for integration purposes (read-only)"
        sendStatus:
          type: "string"
          enum: ["PENDING", "ERROR", "SUCCESS"]
          description: "Status of send operation (read-only)"
        sendNumber:
          type: "string"
          description: "Optional number for integration purposes (read-only)"
        createdAt:
          type: "string"
          format: "date-time"
      required:
        - id
        - business
        - responsible
        - title
        - status
        - date
        - value
        - sendStatus
        - createdAt
      description: "Business proposal entity response"
      example:
        id: "123e4567-e89b-12d3-a456-426614174000"
        business:
          id: "456e7890-e89b-12d3-a456-426614174000"
        responsible:
          id: "789e0123-e89b-12d3-a456-426614174000"
        title: "Software Implementation Proposal"
        status: "Enviado"
        date: "2024-01-15"
        value: 50000
        content: "Detailed proposal content..."
        themeColor: "#007bff"
        termsAndConditions: "Standard terms and conditions..."
        showUnitPrices: true
        sendMessage: null
        sendStatus: "PENDING"
        sendNumber: null
        createdAt: "2024-01-10T08:00:00.000Z"
...
    PaginatedBusinessProposalResponse:
      type: "object"
      properties:
        contents:
          type: "array"
          items:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              business:
                type: "object"
                properties:
                  id:
                    type: "string"
                    format: "uuid"
                required:
                  - id
              responsible:
                type: "object"
                properties:
                  id:
                    type: "string"
                    format: "uuid"
                required:
                  - id
              title:
                type: "string"
              status:
                type: "string"
              date:
                type: "string"
                pattern: "^\\d{4}-\\d{2}-\\d{2}$"
              value:
                type: "number"
                minimum: 0
                exclusiveMinimum: true
              content:
                type: "string"
              themeColor:
                type: "string"
                pattern: "^#[0-9A-Fa-f]{6}$"
              termsAndConditions:
                type: "string"
              showUnitPrices:
                type: "boolean"
              sendMessage:
                type: "string"
                description: "Optional message for integration purposes (read-only)"
              sendStatus:
                type: "string"
                enum: ["PENDING", "ERROR", "SUCCESS"]
                description: "Status of send operation (read-only)"
              sendNumber:
                type: "string"
                description: "Optional number for integration purposes (read-only)"
              createdAt:
                type: "string"
                format: "date-time"
            required:
              - id
              - business
              - responsible
              - title
              - status
              - date
              - value
              - sendStatus
              - createdAt
          description: "Array of BusinessProposal items"
        totalElements:
          type: "integer"
          minimum: 0
          description: "Total number of elements across all pages"
          example: 150
        totalPages:
          type: "integer"
          minimum: 0
          description: "Total number of pages"
          example: 8
      required:
        - contents
        - totalElements
        - totalPages
      description: "Paginated response containing business proposals"
...
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
      description: "JWT token for authentication. Include the token in the Authorization header as: Bearer <token>"
  responses:
    UnauthorizedError:
      description: "Authentication information is missing or invalid"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Unauthorized"
              status:
                type: "integer"
                example: 401
    ValidationError:
      description: "Validation error with field-specific messages"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Validation failed"
              status:
                type: "integer"
                example: 400
              errors:
                type: "array"
                items:
                  type: "object"
                  properties:
                    field:
                      type: "string"
                      example: "name"
                    message:
                      type: "string"
                      example: "Name is required"
    NotFoundError:
      description: "Resource not found"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Resource not found"
              status:
                type: "integer"
                example: 404
...
security:
  - bearerAuth:
...
paths:
...
  /api/business-proposals:
    post:
      operationId: "createBusinessProposal"
      summary: "Create new businessproposal"
      description: "Creates a new businessproposal record in the system"
      tags:
        - Business Proposals
      parameters:
      security:
        - bearerAuth:
      requestBody:
        description: "BusinessProposal data to create"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessProposalRequest"
      responses:
        201:
          description: "BusinessProposal created successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProposal"
        400:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ValidationError"
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/responses/UnauthorizedError"
    get:
      operationId: "getBusinessProposals"
      summary: "Get all businessproposals"
      description: "Retrieves a paginated list of businessproposals with optional filtering\n\n**Example URLs:**\n- **Basic pagination**: `/api/business-proposals?page=1&size=20` - Get first page with 20 items\n- **Custom page size**: `/api/business-proposals?page=2&size=50` - Get second page with 50 items\n- **High-value opportunities**: `/api/business-proposals?filter=value >= 50000 AND stage = \"Negotiation\"` - Get high-value opportunities in negotiation stage\n- **BRL opportunities this year**: `/api/business-proposals?filter=currency = \"BRL\" AND createdAt >= \"2024-01-01\"` - Get BRL opportunities created this year"
      tags:
        - Business Proposals
      parameters:
        - name: "filter"
          in: "query"
          required: false
          schema:
            type: "string"
          description: "Dynamic filter parameter using SQL-like syntax. Supports operators: =, !=, >, <, >=, <=, LIKE, ILIKE, IN, NOT IN, AND, OR. Use parentheses for complex logical grouping."
          example: "status = \"ACTIVE\" AND type = \"Client\""
        - name: "page"
          in: "query"
          required: false
          schema:
            type: "integer"
            minimum: 1
            default: 1
          description: "Page number for pagination (starts from 1)"
          example: 1
        - name: "size"
          in: "query"
          required: false
          schema:
            type: "integer"
            minimum: 1
            maximum: 100
            default: 20
          description: "Number of items per page (maximum 100)"
          example: 20
      security:
        - bearerAuth:
      responses:
        200:
          description: "Paginated list of businessproposals"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedBusinessProposalResponse"
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/responses/UnauthorizedError"
  /api/business-proposals/{id}:
    get:
      operationId: "getBusinessProposalById"
      summary: "Get businessproposal by ID"
      description: "Retrieves a specific businessproposal by its unique identifier\n\n**Example URLs:**\n- **Basic pagination**: `/api/business-proposals/{id}?page=1&size=20` - Get first page with 20 items\n- **Custom page size**: `/api/business-proposals/{id}?page=2&size=50` - Get second page with 50 items\n- **High-value opportunities**: `/api/business-proposals/{id}?filter=value >= 50000 AND stage = \"Negotiation\"` - Get high-value opportunities in negotiation stage\n- **BRL opportunities this year**: `/api/business-proposals/{id}?filter=currency = \"BRL\" AND createdAt >= \"2024-01-01\"` - Get BRL opportunities created this year"
      tags:
        - Business Proposals
      parameters:
        - name: "filter"
          in: "query"
          required: false
          schema:
            type: "string"
          description: "Dynamic filter parameter using SQL-like syntax. Supports operators: =, !=, >, <, >=, <=, LIKE, ILIKE, IN, NOT IN, AND, OR. Use parentheses for complex logical grouping."
          example: "status = \"ACTIVE\" AND type = \"Client\""
        - name: "page"
          in: "query"
          required: false
          schema:
            type: "integer"
            minimum: 1
            default: 1
          description: "Page number for pagination (starts from 1)"
          example: 1
        - name: "size"
          in: "query"
          required: false
          schema:
            type: "integer"
            minimum: 1
            maximum: 100
            default: 20
          description: "Number of items per page (maximum 100)"
          example: 20
      security:
        - bearerAuth:
      responses:
        200:
          description: "BusinessProposal found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProposal"
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/responses/UnauthorizedError"
        404:
          description: "BusinessProposal not found"
          content:
            application/json:
              schema:
                $ref: "#/components/responses/NotFoundError"
    put:
      operationId: "updateBusinessProposal"
      summary: "Update businessproposal"
      description: "Updates an existing businessproposal with new data"
      tags:
        - Business Proposals
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
          description: "BusinessProposal unique identifier"
          example: "123e4567-e89b-12d3-a456-426614174000"
      security:
        - bearerAuth:
      requestBody:
        description: "Updated businessproposal data"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessProposalRequest"
      responses:
        200:
          description: "BusinessProposal updated successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProposal"
        400:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ValidationError"
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/responses/UnauthorizedError"
        404:
          description: "BusinessProposal not found"
          content:
            application/json:
              schema:
                $ref: "#/components/responses/NotFoundError"
    delete:
      operationId: "deleteBusinessProposal"
      summary: "Delete businessproposal"
      description: "Deletes an existing businessproposal from the system"
      tags:
        - Business Proposals
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
          description: "BusinessProposal unique identifier"
          example: "123e4567-e89b-12d3-a456-426614174000"
      security:
        - bearerAuth:
      responses:
        204:
          description: "BusinessProposal deleted successfully"
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/responses/UnauthorizedError"
        404:
          description: "BusinessProposal not found"
          content:
            application/json:
              schema:
                $ref: "#/components/responses/NotFoundError"
...